pyproject.toml

[project]
name = "memory-service"
version = "0.1.0"
description = "Personal Memory MCP Service - A memory management service using LlamaIndex and FastMCP"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
{name = "victor", email = "victorwtn@outlook.com"}
]
dependencies = [
    "fastmcp>=0.1.0",
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-openai-like",
    "llama-index-llms-openai-like>=0.3.5",
    "llama-index-readers-json>=0.3.0",
    "llama-index-storage-docstore-postgres>=0.3.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "httpx>=0.28.1",
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
"pytest>=7.0.0",
"black>=23.0.0",
"isort>=5.12.0",
"mypy>=1.0.0",
"ruff>=0.0.270",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["memserv"]

[tool.pytest.ini_options] 
pythonpath = [
".", "src",
]


[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "B", "I"]
ignore = []

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
"pytest>=8.4.1",
]
